[➕ Add] Equip Update
