--- GoodFunGame/Assets/Scripts/UI/SubItem/UI_SkillEquipButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7de9dacf9e5f0784c98db355e9e67c4e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
